Add an exception in case of proc_open() failure
When proc_open() does not return a valid resource the run method does nothing, and $this->exitCode is still set to null. The result is that AbstractGenerator::checkProcessStatus() throws an exception with a very mysterious message :

The exit status code '' says something went wrong:
stderr: ""
stdout: ""
command: wkhtmltopdf [...]

So throwing a RuntimeException earlier would be a good idea, like it's done in symfony/Process.