Fix memory leak
Calling `register_shutdown_function` with a callback referencing `$this`
means `$this` needs to stay in memory until the end of the program, the
garbage collector can't free the memory until after the shutdown
function is called. In short lived scripts the impact is very small, but
in large test suite where the Generator is instantiated for every test
(i.e. in the symfony container) the impact can be measured in hundred of
megabytes or even in gigabytes.

To fix it we avoid referencing `$this` by calling a static function
instead. We keep a separate list of temporary files to clean in a static
list. Each Generator will clean its own temporary files on `__destruct()`
and any remaining files will be cleaned by the shutdown function.